Add statement from Baumgartner's proof of Erdos question.
Also adding context around Erdos 198.

This was a bit confusing at first because Bloom references a result in the opposite direction to what he states and it seems (according to Erdos and Graham) that the same author proved results in each direction, but a reference for only one direction is provided.

PiperOrigin-RevId: 731324074
Change-Id: I7a324d14b66151abb1de01f787a67887f506ff0f