feat: add answer(sorry) to all Erdos problems stated as questions. (#68)
See #29.

This fixes this for all open Erdos problems, and also for the solved
ones up to 442.

Also fixes a potential misformalization / issue with 244, where
previously it was assumed that the set had some density (which does not
seem clear from the problem statement).

Also, when creating this branch I by accident also had some work in
converting (a few) docstrings into Latex. I have not bothered to remove
this, as this is a (very minor) good thing to do anyways.

---------

Co-authored-by: Paul Lezeau <lezeau@google.com>