Local connectivity of the Mandelbrot and Multibrot sets (#112)
Adds the [MLC
conjecture](https://en.wikipedia.org/wiki/Mandelbrot_set#Local_connectivity),
stating that the Mandelbrot set is locally connected / that all
Multibrot sets are. This is the first of the two conjectures from #92.

I'm somewhat unsure whether this should be placed in the wikipedia
folder or in the folder for a paper; the linked wikipedia article
mentions the conjecture only for the Mandelbrot set, while e.g. [this
paper](https://arxiv.org/abs/math/9902155) talks about the conjecture
for Multibrot sets too.

---------

Co-authored-by: Paul Lezeau <paul.lezeau@gmail.com>
Co-authored-by: Paul Lezeau <lezeau@google.com>