Add Modularity Conjecture  (#168)
### What is the conjecture
State the Shimura--Taniyama--Weil conjecture (also known as the
modularity conjecture/theorem) for elliptic curves over the rational
numbers.

https://en.wikipedia.org/wiki/Modularity_theorem

This is a very hard, known theorem in the case of the rational number
and many other cases, but open in general (and much harder to state!).
The version we state is stronger than what is needed for FLT.

### [AMS
categories](https://github.com/google-deepmind/formal-conjectures/labels?q=ams-)

* ams-11

---------

Co-authored-by: Calle Sönne <calle.sonne@gmail.com>