chore(ErdosProblems/402): fix misformalisation (#263)
Reformulate the equality case.
- The equality case is intended to assert that the mutual gcd is 1: this
is on the website and in the docstring already
- As a consequence, no assumption on the cardinality is needed (eg for
singletons, we can only have the gcd being 1 if it's already {1}, and
then we're in the first of the three cases, so the theorem holds).
- The equality case is stated here by saying that (a,b) >= a/|A| for all
a,b, as written in the fourth paragraph of the introduction in [BaSo96].