add theorem for $\pi^{\pi^{\pi^\pi}}$ an integer (#607)
This PR fixes #592.

It adds the conjecture that $\pi^{\pi^{\pi^\pi}}$ is not an integer.

This is similar to questions already in the `Transcendental` file, and
would follow from $\pi^{\pi^{\pi^\pi}}$ being transcendental (which I
have taken the liberty of adding as well), so we include it there.

But this formulation about being an integer is of interest in its own
right, as it could in principle be proven by direct computation. See
this [YouTube Video](https://www.youtube.com/watch?v=BdHFLfv-ThQ) for
more discussion.