Erdős Problem 723 (#612)
Adds [Erdős Problem 723](https://www.erdosproblems.com/723).

I'm not sure if this is actually due to Erdős. This seems to be the
Prime Power Conjecture. I looked through the linked paper and I couldn't
find it. I think it might be better to move the contents to
`FormalConjectures/Other/PrimePower.lean` and import it into
`FormalConjectures/ErdosProblems/723.lean`, similar to the Brocard
Problem.