chore: bump toolchain and Mathlib to v4.18.0 (#111)
This PR updates the Lean toolchain and Mathlib dependency to v4.18.0.

This requires two additional changes:
* The Mathlib module `Mathlib.Logic.Equiv.Fin` has been split, and the
newly relevant import is `Mathlib.Logic.Equiv.Fin.Rotate`. I've also
update the name of the relevant module in `ForMathlib/`
* A golfed proof by AlphaProof submitted in
https://github.com/kim-em/formal-conjectures/commit/8ec658213ab9797e3f7fb4a1fe0e2bda37307897
actually made the proof more fragile, and in particular more fragile in
way that broke when updating the toolchain. For now, I have reverted the
proof for simplicity.

I'd encourage anyone interested to re-golf that proof in a way that
remains compatible with v4.18.0.

---------

Co-authored-by: Moritz Firsching <firsching@google.com>
Co-authored-by: Bhavik Mehta <bhavikmehta8@gmail.com>